Revision 571:
Bug fix:
-avoid following target after restart-logout , thanks to EXCLUS1VE for
report.
-pick up stuck , thanks to EXCLUS1VE for report.
-summon patk/matk speed visual animation.
-summon attack request , thanks to EXCLUS1VE for report.
-effect relax , thanks to EXCLUS1VE for report.
-formulas hitmiss , thanks to EXCLUS1VE for report.
-Social action  request , thanks to EXCLUS1VE for report.
-cubic attack , thanks to EXCLUS1VE for report.
-Party match room chat, thanks to EXCLUS1VE for report.
-pick up issue , thanks to EXCLUS1VE for report.
-pet npe, thanks to EXCLUS1VE for report.


Rework:
-player template.
-skilltreedata.
-RequestAquireSkill.
-RequestAquireSkillInfo.
-RequestExEnchantSkill.
-RequestExEnchantSkillInfo.
-AcquireSkillList.
-ExEnchantSkillList.

Optimize:
-players got their own getMoveSpeed getter , and triggers when speed
change by user(walk/run commands) or buff-debuff,
that way we avoid unnecessary speed calculation by updateposition task
that literaly spam the calculators.
-castle traps are optimized and enabled.

damage calculation formula:
-blow damage position bonus.
-when you make a magic critical hit, the magic damage is now tripled.
-melee normal attack Damage position bonus.

Implement:
-new vote api system.(npc-> //spawn 61)

Organization:
FenceManager moved in -> datatables/xml and rename to FenceData. 

Remove:
-old antibot system and kept only one as main.
-old vote system.

+++ L2JHellasC/java/Extensions/fake/roboto/FakePlayer.java
@@ -235,14 +235,14 @@ public class FakePlayer extends L2PcInstance
 			{
 				if (sklTargetType == L2SkillTargetType.TARGET_GROUND)
 				{
-					if (!isInsideRadius(worldPosition.getX(), worldPosition.getY(), worldPosition.getZ(), skill.getCastRange() + getTemplate().getCollisionRadius(), false, false))
+					if (!isInsideRadius(worldPosition.getX(), worldPosition.getY(), worldPosition.getZ(), skill.getCastRange() + (int)getTemplate().getCollisionRadius(), false, false))
 					{
 						sendPacket(SystemMessageId.TARGET_TOO_FAR);
 						sendPacket(ActionFailed.STATIC_PACKET);
 						return false;
 					}
 				}
-				else if (skill.getCastRange() > 0 && !isInsideRadius(target, skill.getCastRange() + getTemplate().getCollisionRadius(), false, false))
+				else if (skill.getCastRange() > 0 && !isInsideRadius(target, skill.getCastRange() + (int)getTemplate().getCollisionRadius(), false, false))
 				{
 					sendPacket(SystemMessageId.TARGET_TOO_FAR);
 					sendPacket(ActionFailed.STATIC_PACKET);
